refactor: move controller methods to the endpoint class
